Fixed on/off button.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/lib/cf.h
@@ -123,8 +123,8 @@ namespace K {
         Isolate* isolate = args.GetIsolate();
         HandleScope scope(isolate);
         Local<Object> o = Object::New(isolate);
-        o->Set(FN::v8S("base"), Number::New(isolate, (double)cfBase()));
-        o->Set(FN::v8S("quote"), Number::New(isolate, (double)cfQuote()));
+        o->Set(FN::v8S("base"), Number::New(isolate, cfBase()));
+        o->Set(FN::v8S("quote"), Number::New(isolate, cfQuote()));
         args.GetReturnValue().Set(o);
       };
   };
